rename: src폴더 내에 파일 MVC패턴에 맞춰 변경
