+++ src/Controllers/Validator/purchasePriceValidator.js
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGE } from '../Constants/errorMessages.js';
+import { ERROR_MESSAGE } from '../../Constants/errorMessages.js';
 
 const isPurchasePriceEmptyInput = (purchasePriceInput) => purchasePriceInput !== '';
