+++ src/Models/winningNumbersUtils.js
@@ -1,5 +1,6 @@
 const trimWinningNumbers = (winningNum) => {
   const trimWinningNum = winningNum.split(',').map((x) => x.trim());
+  
   return trimWinningNum;
 };
